rename: 컴포넌트 첫글자를 대문자로 수정
