+++ src/app/_components/common/Popup.tsx
@@ -1,7 +1,7 @@
 import { paragraph2, semiBold } from '@/app/_styles/font.css';
 import { flexSprinklesFc } from './utils/flex';
 import { popupContents, popupWrapper } from './popup.css';
-import { usePopupStore } from '@/app/_store/popup/PopupStore';
+import { usePopupStore } from '@/app/_store/popup/popupStore';
 
 const Popup = ({ children }: { children: React.ReactNode }) => {
   const message = usePopupStore((state) => state.message);
