+++ src/app/edit/[date]/[id]/layout.tsx
@@ -1,5 +1,7 @@
 import { recordContainer } from '@/app/record/_styles/record.css';
 
-export default function Layout({ children }: { children: React.ReactNode }) {
+const Layout = ({ children }: { children: React.ReactNode }) => {
   return <section className={recordContainer}>{children}</section>;
-}
+};
+
+export default Layout;
