+++ src/app/graph/_components/dataGraph.tsx
@@ -14,7 +14,7 @@ import {
   ChartData,
 } from 'chart.js';
 import { flexSprinklesFc } from '@/app/_components/common/utils/flex';
-import { filterWrapper, toggle, toggleActive } from '../_styles/graph.css';
+import { chartBg, filterWrapper, toggle, toggleActive } from '../_styles/graph.css';
 import Image from 'next/image';
 import { useEffect, useState } from 'react';
 import { caption, heading2, regular, semiBold } from '@/app/_styles/font.css';
@@ -22,7 +22,11 @@ import { pointer } from '@/app/_styles/global.css';
 import { paddingSprinkles } from '@/app/_styles/padding.css';
 import { supabaseClient } from '@/app/_lib/supabaseClient';
 import { transformBowelData, TransformedBowelData } from '../_utils';
-import ShowDataGraph from './showDataGraph';
+import { Line } from 'react-chartjs-2';
+import { CustomLineGraph } from '../_plugins/chart/CustomLineGraph';
+import PoopInfoBox from './PoopInfoBox';
+import { useBowelData } from '../_hook/useBowelData';
+import BowelChartContainer from './BowelChartContainer';
 
 ChartJS.register(
   CategoryScale,
@@ -86,12 +90,18 @@ const DataGraph = () => {
           onClick={() => setIsToggleActive(!isToggleActive)}
         />
       </div>
-      <ShowDataGraph
-        hasData={hasData}
-        dateRange={dateRange}
-        bowelDate={bowelDate}
-        isToggleActive={isToggleActive}
-      />
+      <>
+        {hasData ? (
+          <BowelChartContainer bowelDate={bowelDate} isToggleActive={isToggleActive} />
+        ) : (
+          <>
+            <p className={paddingSprinkles({ paddingTop: 's28' })}>
+              {dateRange === 7 ? '일주일' : '한 달'} 동안의 배변 기록이 없습니다. 😥
+            </p>
+            <p>새로운 기록을 추가해보세요.</p>
+          </>
+        )}
+      </>
     </>
   );
 };
