+++ src/app/my/page.tsx
@@ -1,11 +1,11 @@
 'use client';
 
-import UserProfileImage from './_components/userProfileImage';
-import UserGoalInput from './_components/userGoalInput';
-import LogoutButton from './_components/logoutButton';
-import MyPopup from './_components/myPopup';
+import UserProfileImage from './_components/UserProfileImage';
+import UserGoalInput from './_components/UserGoalInput';
+import LogoutButton from './_components/LogoutButton';
+import MyPopup from './_components/MyPopup';
 
-const Page = () => {
+const MyPage = () => {
   return (
     <>
       <MyPopup />
@@ -19,4 +19,4 @@ const Page = () => {
   );
 };
 
-export default Page;
+export default MyPage;
