+++ src/app/record/(info)/detail/_components/DetailPopup.tsx
@@ -1,6 +1,6 @@
 import Button from '@/app/_components/common/Button';
 import Popup from '@/app/_components/common/Popup';
-import { usePopupStore } from '@/app/_store/popup/PopupStore';
+import { usePopupStore } from '@/app/_store/popup/popupStore';
 import { colors } from '@/app/_styles/colors.css';
 
 const DetailPopup = () => {
