+++ src/app/record/(info)/detail/_components/TitleText.tsx
