+++ src/app/record/(info)/detail/page.tsx
@@ -6,12 +6,12 @@ import { flexSprinklesFc } from '@/app/_components/common/utils/flex';
 import { colors } from '@/app/_styles/colors.css';
 
 import useInfoStore from '@/app/_store/info/infoStore';
-import { usePopupStore } from '@/app/_store/popup/PopupStore';
-import DetailPopup from './components/popup';
-import TitleText from './components/titleText';
+import { usePopupStore } from '@/app/_store/popup/popupStore';
+import DetailPopup from './_components/DetailPopup';
 import { infoContainer } from '../_common/common.css';
 import { Step, StepChangeHandler } from '../../page';
-import useDetailActions from './hook/useDetailActions';
+import useDetailActions from './_hook/useDetailActions';
+import TitleText from './_components/titleText';
 
 const DetailPage = ({ onButtonClick }: { onButtonClick: StepChangeHandler }) => {
   const detailPopupState = usePopupStore((state) => state.openPopup);
