+++ src/app/record/(info)/shape/_styles/shape.css.ts
