+++ src/app/record/index.tsx
@@ -5,15 +5,15 @@ import Memo from '../_components/common/Memo';
 import { flexSprinklesFc } from '../_components/common/utils/flex';
 import useInfoStore from '../_store/info/infoStore';
 
-import RecordPopup from './_components/popup';
-import RecordCalender from './_components/recordCalender';
+import RecordPopup from './_components/RecordPopup';
+import RecordCalender from './_components/RecordCalender';
 import { plusIconBox, plusIcon } from './_styles/record.css';
 import Image from 'next/image';
 import { Step, StepChangeHandler } from './page';
 import { supabaseClient } from '../_lib/supabaseClient';
 import { BowelAttributes } from '../_types/bowelAttributesSchema';
 import { useUserInfoStore } from '../_store/user/userStore';
-import { usePopupStore } from '../_store/popup/PopupStore';
+import { usePopupStore } from '../_store/popup/popupStore';
 
 const RecordPage = ({ onButtonClick }: { onButtonClick: StepChangeHandler }) => {
   const [filteredData, setFilteredData] = useState<BowelAttributes[] | []>([]);
